pyproject.toml
[tool.poetry]
name = "lightpipe"
version = "0.1.0"
description = "LlamaIndex + LightRAG Pipeline in python"
authors = ["The AI Product Guy <aiproductguy@pm.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.111.0"
uvicorn = {extras = ["standard"], version = "0.22.0"}
pydantic = "2.7.1"
python-multipart = "0.0.9"
python-socketio = "*"
grpcio = "*"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
PyJWT = {extras = ["crypto"], version = "*"}
requests = "2.32.2"
aiohttp = "3.9.5"
httpx = "*"
openai = "*"
anthropic = "*"
google-generativeai = "*"
vertexai = "*"
pymongo = "*"
peewee = "*"
SQLAlchemy = "*"
boto3 = "*"
redis = "*"
sqlmodel = "*"
chromadb = "*"
psycopg2-binary = "*"
langfuse = "*"
ddtrace = "*"
torch = "*"
numpy = "*"
pandas = "*"
xgboost = "*"
scikit-learn = "*"
sentence-transformers = "*"
transformers = "*"
tokenizers = "*"
nltk = "*"
tiktoken = "*"
Pillow = "*"
opencv-python = "*"
matplotlib = "*"
seaborn = "*"
selenium = "*"
playwright = "*"
beautifulsoup4 = "*"
llama-index = "*"
llama-index-llms-ollama = "*"
starlette = "*"
sseclient-py = "^1.8.0"
lightrag-hku = "^1.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
